+++ src/kernel/tests/pmm_tests.cpp
@@ -1,3 +1,4 @@
+#include <cstdint>
 #include <tests/ktest.hpp>
 #include <memory/pmm.hpp>
 #include <memory/vmm.hpp>
@@ -9,11 +10,11 @@ namespace Tests {
         KTest::reset();
 
         // alloc returns a non-zero physical address
-        std::uintptr_t pa1 = Memory::Pmm::alloc(1);
+        std::uint64_t pa1 = Memory::Pmm::alloc(1);
         KTEST_ASSERT(pa1 != 0, "alloc(1) returns non-zero");
 
         // alloc returns a second different address
-        std::uintptr_t pa2 = Memory::Pmm::alloc(1);
+        std::uint64_t pa2 = Memory::Pmm::alloc(1);
         KTEST_ASSERT(pa2 != 0,   "alloc(1) second call returns non-zero");
         KTEST_ASSERT(pa2 != pa1, "alloc(1) returns unique addresses");
 
@@ -23,7 +24,7 @@ namespace Tests {
         KTEST_ASSERT(*ptr == 0xCAFEBABEull, "allocated page is writable via HHDM");
 
         // multi-page alloc
-        std::uintptr_t pa3 = Memory::Pmm::alloc(4);
+        std::uint64_t pa3 = Memory::Pmm::alloc(4);
         KTEST_ASSERT(pa3 != 0, "alloc(4) returns non-zero");
 
         // free returns pages to pool — next alloc should succeed
@@ -31,12 +32,12 @@ namespace Tests {
         Memory::Pmm::free(pa2, 1);
         Memory::Pmm::free(pa3, 4);
 
-        std::uintptr_t pa4 = Memory::Pmm::alloc(1);
+        std::uint64_t pa4 = Memory::Pmm::alloc(1);
         KTEST_ASSERT(pa4 != 0, "alloc after free returns non-zero");
         Memory::Pmm::free(pa4, 1);
 
         // pages are 4KB aligned
-        std::uintptr_t pa5 = Memory::Pmm::alloc(1);
+        std::uint64_t pa5 = Memory::Pmm::alloc(1);
         KTEST_ASSERT((pa5 & 0xFFF) == 0, "alloc returns page-aligned address");
         Memory::Pmm::free(pa5, 1);
